+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # kari CHANGELOG
 
+### 0.2.0
+* Add math methods:
+  * add
+  * divide
+  * modulo (mathematical modulo)
+  * multiply
+  * remainder (JS modulo)
+  * subtract
+
 ### 0.1.2
 * Fix explanation of placeholders in readme